refactor: Improve configurations
The out of the box operator sdk is a bit cumbersome, I had to create a separate environments for dev and prod which IMO should have been the default behavior of this SDK. It's a common thing, you develop an Operator you want it to work in dev and prod slightly different.

Also introduce prettier for formatting the YAMLs in a consistent way after they are being generated from the Go lang docblocks - honestly I'm not happy about this approach of kubebuilder generating manifests from code instead of code from manifests, it's not following the DDD approach where the domain doesn't need to be for a specific language. But for now I'll just adjust it slightly not too much.

- Created a new Kustomization file for metrics configuration.
- Added a Service resource for the inference gateway metrics.
- Updated the network policy to allow metrics traffic from namespaces labeled with 'metrics: enabled'.
- Modified existing Kustomization files to include new resources.
- Enhanced RBAC roles to manage access to metrics and gateway resources.
- Updated the gateway controller to handle new environment variables and container ports.
- Added additional printer columns to the Gateway CRD for better visibility.
- Removed the deprecated namespace installation manifest.

Signed-off-by: Eden Reich <eden.reich@gmail.com>