feat: Enhance Ingress configuration with host and TLS options (#3)
* feat: Enhance Ingress configuration with host and TLS options

Start from simplest configuration option to advance, in case people wants it fully configurable. Simplest approach works for most use-cases - where you configure a cert-manager and a cluster-issuer and just reference it.

Signed-off-by: Eden Reich <eden.reich@gmail.com>

* chore: Add additional API keys to gateway secret configurations

Signed-off-by: Eden Reich <eden.reich@gmail.com>

* refactor: Improve configurations

The out of the box operator sdk is a bit cumbersome, I had to create a separate environments for dev and prod which IMO should have been the default behavior of this SDK. It's a common thing, you develop an Operator you want it to work in dev and prod slightly different.

Also introduce prettier for formatting the YAMLs in a consistent way after they are being generated from the Go lang docblocks - honestly I'm not happy about this approach of kubebuilder generating manifests from code instead of code from manifests, it's not following the DDD approach where the domain doesn't need to be for a specific language. But for now I'll just adjust it slightly not too much.

- Created a new Kustomization file for metrics configuration.
- Added a Service resource for the inference gateway metrics.
- Updated the network policy to allow metrics traffic from namespaces labeled with 'metrics: enabled'.
- Modified existing Kustomization files to include new resources.
- Enhanced RBAC roles to manage access to metrics and gateway resources.
- Updated the gateway controller to handle new environment variables and container ports.
- Added additional printer columns to the Gateway CRD for better visibility.
- Removed the deprecated namespace installation manifest.

Signed-off-by: Eden Reich <eden.reich@gmail.com>

* refactor: Simplify test task dependencies in Taskfile

Signed-off-by: Eden Reich <eden.reich@gmail.com>

* refactor: Remove unnecessary task dependencies from e2e test workflow - keep it simple

Signed-off-by: Eden Reich <eden.reich@gmail.com>

* refactor: Removing metrics checks for now, will improve it on another iteration

I want to breakdown the entire e2e testing file to smaller digestable chunks split by domain.
For example all Ingress related tests in e2e_ingress etc..

Signed-off-by: Eden Reich <eden.reich@gmail.com>

* refactor: Remove unnecessary dependencies from install and uninstall tasks

Signed-off-by: Eden Reich <eden.reich@gmail.com>

* chore: Update config/crd/bases/core.inference-gateway.com_gateways.yaml

---------

Signed-off-by: Eden Reich <eden.reich@gmail.com>