fix: Update gateway configuration for development environment and disable certain features
Decided to go with a more Kubernetes native approach by using rolling updates to load the new environment variables - it's simple and for a reason.
I think the reload of config using SIGHUP is not really necessary, I just need to ensure that there is a proper graceful shutdown in the application because clients going to have SSE persistent connections, which means I would need to gracefully wait for those connection to complete or timeout before shutting down an instance.

Signed-off-by: Eden Reich <eden.reich@gmail.com>