refactor: Improve the overall configurations experience of the Gateway (#9)
* fix: Update inference-gateway image version and clean up unused labels and configurations

Make this example less confusing.

Signed-off-by: Eden Reich <eden.reich@gmail.com>

* fix: Update gateway configuration for development environment and disable certain features

Decided to go with a more Kubernetes native approach by using rolling updates to load the new environment variables - it's simple and for a reason.
I think the reload of config using SIGHUP is not really necessary, I just need to ensure that there is a proper graceful shutdown in the application because clients going to have SSE persistent connections, which means I would need to gracefully wait for those connection to complete or timeout before shutting down an instance.

Signed-off-by: Eden Reich <eden.reich@gmail.com>

* feat: Add URL field for provider API endpoints and update related configurations

Signed-off-by: Eden Reich <eden.reich@gmail.com>

* fix: Ensure API key is only accessed if SecretRef is not nil

Signed-off-by: Eden Reich <eden.reich@gmail.com>

* feat: Refactor ProviderSpec to use environment variables instead of secret references for API keys

Maybe it looks a bit verbose, but this will ensure rolling update and also it gives the user more flexibility in terms of how they want the keys and in the secret to be called.

Signed-off-by: Eden Reich <eden.reich@gmail.com>

* test: Update Gateway controller to simplify provider API key handling in tests

Signed-off-by: Eden Reich <eden.reich@gmail.com>

* test: Remove ConfigMap reconciliation and update tests to reflect changes

There is no longer configmap owned by the controller, it just allows the user to define it and reference it but it's not managing it.

Signed-off-by: Eden Reich <eden.reich@gmail.com>

* test: Remove OTLP gRPC endpoint validation from telemetry configuration tests

Since it's not implemented yet by the inference gateway it's not relevant.

The gateway has prometheus exporter and an endpoint at /metrics

Traces are currently not being pushed to external otel collector using OTLP grpc, future implementation perhaps if needed.

Signed-off-by: Eden Reich <eden.reich@gmail.com>

* refactor: Add nolint directive to buildContainer function for cyclomatic complexity

Will refactor it later, it's not important right now.

Signed-off-by: Eden Reich <eden.reich@gmail.com>

* test: Update telemetry configuration to disable by default and remove tracing support

Signed-off-by: Eden Reich <eden.reich@gmail.com>

* test: Refactor Gateway controller tests to improve telemetry configuration validation

Also change the default value of TELEMETRY to false, it should not be enabled unless the user specified true

Signed-off-by: Eden Reich <eden.reich@gmail.com>

* refactor: Simplify telemetry setting determination in Gateway reconciler

Signed-off-by: Eden Reich <eden.reich@gmail.com>

* test: Update OpenTelemetry configuration tests to remove tracing support

Signed-off-by: Eden Reich <eden.reich@gmail.com>

* refactor: Enhance HPA and TLS Secret configurations in install.yaml

Use the actual kubernetes autoscaling HPA spec, so we don't need to modify too much in the future unless Kubernetes come up with a new autoscaling API.
Also it helps because people already used to this API so why we should remove it.

- Updated HPA configuration to include detailed scaling behavior settings, including scaleUp and scaleDown policies, stabilization windows, and metrics specifications.
- Introduced new properties for HPA metrics, allowing for more granular control over scaling based on resource utilization and external metrics.
- Improved documentation for secret references in TLS configurations, clarifying the requirements for secret names and keys.
- Added optional fields to specify whether the Secret or its key must be defined, enhancing flexibility in configuration.
- Set default values for metrics configuration to streamline setup.

Signed-off-by: Eden Reich <eden.reich@gmail.com>

* chore: Add todo

Signed-off-by: Eden Reich <eden.reich@gmail.com>

* feat: Implement Custom HPA configuration for Gateway

Start with minimal configurations with defaults and if the user wants they can use explicit config.

Signed-off-by: Eden Reich <eden.reich@gmail.com>

* test: Update HPA tests to reflect new configuration structure and add default value verification

Signed-off-by: Eden Reich <eden.reich@gmail.com>

---------

Signed-off-by: Eden Reich <eden.reich@gmail.com>