feat: Refactor A2A resource to use Card structure and SkillsList, update CRD and controller logic
Add the skills to the table only if -o wide specified as a comma separated list. The list could get sometimes too long, not sure if that's a good idea, but will keep like this for now, perhaps will change it in the future and we will need to use kubectl describe a2a to see the card.

Signed-off-by: Eden Reich <eden.reich@gmail.com>