feat: Add MCP server configuration with namespace and specifications
Would be cool to just have a side-car container that ensures those STDIO MCP servers can also support SSE or any streamable HTTP.

Some MCPs supports only STDIO. which makes it complex to connect to in a containerized environment. Meaning you would normally have to create a custom bridge (web server) that sends a command to the internal process - not ideal, but it-is what it-is.

Signed-off-by: Eden Reich <eden.reich@gmail.com>